housekeeping the ActiveSpanSource in a separated module
