+++ proposal/tracer.py
@@ -1,5 +1,7 @@
 from opentracing.tracer import Tracer as OTBaseTracer
 
+from .active_span_source import NoopActiveSpanSource
+
 
 class Tracer(OTBaseTracer):
     """Tracer class that provides some extension to the base OpenTracing
@@ -35,27 +37,3 @@ class Tracer(OTBaseTracer):
         # set it as active Span
         self.active_span_source.make_active(span)
         return span
-
-
-class BaseActiveSpanSource(object):
-    """BaseActiveSpanSource is the interface for a pluggable class that
-    defines the source for the ActiveSpan. It must be used as a base class
-    to implement the right behavior in the vendors specific implementation.
-    """
-    def make_active(self, span):
-        raise NotImplementedError
-
-    def active_span(self, span):
-        raise NotImplementedError
-
-
-class NoopActiveSpanSource(BaseActiveSpanSource):
-    """ActiveSpanSource provides the logic to get and set the current
-    ActiveSpan. This is a cheap noop implementation that must be implemented
-    by Tracer developers.
-    """
-    def make_active(self, span):
-        return None
-
-    def active_span(self):
-        return None
